signals folder norm complient
