+++ src/Signals/signals.c
@@ -6,7 +6,7 @@
 /*   By: rmakoni <rmakoni@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 12:26:06 by ksinn             #+#    #+#             */
-/*   Updated: 2025/04/22 12:52:06 by rmakoni          ###   ########.fr       */
+/*   Updated: 2025/04/24 13:37:19 by rmakoni          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,27 +56,6 @@ static void	handle_sigquit_noninteractive(int signum)
 }
 
 /**
- * @brief Sets up signal handlers for interactive mode
- *
- * This function configures SIGINT to be handled by handle_sigint_interactive
- * and ignores SIGQUIT.
- */
-void	setup_interactive_signals(void)
-{
-	struct sigaction	sa_int;
-	struct sigaction	sa_quit;
-
-	sa_int.sa_handler = handle_sigint_interactive;
-	sa_int.sa_flags = 0;
-	sigemptyset(&sa_int.sa_mask);
-	sigaction(SIGINT, &sa_int, NULL);
-	sa_quit.sa_handler = SIG_IGN;
-	sa_quit.sa_flags = 0;
-	sigemptyset(&sa_quit.sa_mask);
-	sigaction(SIGQUIT, &sa_quit, NULL);
-}
-
-/**
  * @brief Sets up signal handlers for non-interactive mode
  *
  * This function configures SIGINT to be handled by handle_sigint_noninteractive
@@ -98,33 +77,22 @@ void	setup_noninteractive_signals(void)
 }
 
 /**
- * @brief Resets signal handlers to their default behavior
+ * @brief Sets up signal handlers for interactive mode
  *
- * This function restores the default handlers for SIGINT and SIGQUIT.
+ * This function configures SIGINT to be handled by handle_sigint_interactive
+ * and ignores SIGQUIT.
  */
-void	reset_signals(void)
+void	setup_interactive_signals(void)
 {
-	signal(SIGINT, SIG_DFL);
-	signal(SIGQUIT, SIG_DFL);
-}
+	struct sigaction	sa_int;
+	struct sigaction	sa_quit;
 
-/**
- * @brief Retrieves the status code based on the last received signal
- * @return The status code: 130 for SIGINT, 131 for SIGQUIT, or 0 otherwise
- *
- * This function checks the global signal flag, determines the appropriate
- * status code, and resets the flag.
- */
-int	get_signal_status(void)
-{
-	int	status;
-
-	if (g_signal_received == SIGINT)
-		status = 130;
-	else if (g_signal_received == SIGQUIT)
-		status = 131;
-	else
-		status = 0;
-	g_signal_received = 0;
-	return (status);
+	sa_int.sa_handler = handle_sigint_interactive;
+	sa_int.sa_flags = 0;
+	sigemptyset(&sa_int.sa_mask);
+	sigaction(SIGINT, &sa_int, NULL);
+	sa_quit.sa_handler = SIG_IGN;
+	sa_quit.sa_flags = 0;
+	sigemptyset(&sa_quit.sa_mask);
+	sigaction(SIGQUIT, &sa_quit, NULL);
 }
